refactor to use Outlet
